backend/requirements.txt
# ===================================================================
# SECURE ACADEMIC PROJECT SUBMISSION & VERIFICATION PORTAL
# Python Dependencies
# ===================================================================

# Web Framework
Flask==3.0.0
flask-cors==4.0.0

# JWT Authentication
PyJWT==2.8.0

# Cryptography Libraries (Industry Standard)
pycryptodome==3.19.0

# Database
# SQLite is included in Python standard library, no additional package needed

# For production deployment (optional):
# gunicorn==21.2.0
# python-dotenv==1.0.0

# ===================================================================
# Installation Instructions:
# ===================================================================
# 
# 1. Create a virtual environment (recommended):
#    python -m venv venv
#    
# 2. Activate virtual environment:
#    - Windows: venv\Scripts\activate
#    - Linux/Mac: source venv/bin/activate
#    
# 3. Install dependencies:
#    pip install -r requirements.txt
#
# ===================================================================
